24th commit

+++ src/pages/Cart/Cart.tsx
@@ -1,47 +1,49 @@
+/* eslint-disable prefer-const */
+/* eslint-disable @typescript-eslint/no-explicit-any */
 import {
   deleteCartItems,
   getAllCartItems,
   updateCartQuantity,
 } from "@/redux/features/cart/cartSlice";
 import { useAppDispatch, useAppSelector } from "@/redux/hooks";
-import React, { useEffect, useState } from "react";
+import React, { useEffect, useRef, useState } from "react";
 import { Flex, Input, Spin, Table } from "antd";
 import type { TableColumnsType } from "antd";
 import { Button } from "@/components/ui/button";
 import { RiDeleteBin6Fill } from "react-icons/ri";
 import { useNavigate, useOutletContext } from "react-router-dom";
-import img from "../../assets/Result/no-data-found.png";
+import img from "../../assets/Result/empty_cart.png";
 import { Space } from "antd";
 import Swal from "sweetalert2";
 import { useGetallProductsQuery } from "@/redux/features/product/productApi";
 import { LoadingOutlined } from "@ant-design/icons";
+import { TableRowSelection } from "antd/es/table/interface";
 
 interface DataType {
   key: React.Key;
-  _id?: string;
-  name: string;
-  price: number;
-  image: string;
-  quantity: number;
-  quantityInStock?: number;
+  _id?: string | null;
+  name: string | null;
+  price: number | null;
+  image: string | null;
+  quantity: number | null;
+  quantityInStock?: number | null;
 }
 
 const Cart = () => {
   const [selectedRowKeys, setSelectedRowKeys] = useState<React.Key[]>([]);
   const [showMultipleDeleteButton, setShowMultipleDeleteButton] =
     useState(false);
-  const [loading, setLoading] = useState(false);
 
-  const {
-    selectedCartItems,
-    setSelectedCartItems,
-    disabledCartButtons,
-    setDisabledCartButtons,
-  } = useOutletContext();
+  const { selectedCartItems, setSelectedCartItems, setDisabledCartButtons } =
+    useOutletContext<any>();
 
   const navigate = useNavigate();
   const dispatch = useAppDispatch();
 
+  // Refs to store previous values
+  const prevCartItemsRef = useRef<any[]>([]);
+
+  // get the existing products in DB and cart
   const cartItems = useAppSelector(getAllCartItems);
   const {
     data: allProducts,
@@ -49,16 +51,32 @@ const Cart = () => {
     isError: isAllProductsError,
   } = useGetallProductsQuery({ limit: 5000 });
 
+  // set the cart items in table data
   const data: DataType[] = cartItems.map((item, index) => ({
     key: index,
-    _id: item._id,
-    image: item.image,
-    name: item.name,
-    quantity: item.quantity,
-    quantityInStock: item.quantityInStock,
-    price: item.price,
+    _id: item._id ?? null,
+    image: item.image ?? null,
+    name: item.name ?? null,
+    quantity: item.quantity ?? null,
+    quantityInStock: item.quantityInStock ?? null,
+    price: item.price ?? null,
   }));
 
+  // synchronize selectedRowKeys with selectedCartItems
+  useEffect(() => {
+    if (cartItems !== prevCartItemsRef.current) {
+      const selectedKeys = data
+        .filter((item) =>
+          selectedCartItems.some((selected: any) => selected._id === item._id)
+        )
+        .map((item) => item.key);
+
+      setSelectedRowKeys(selectedKeys);
+      prevCartItemsRef.current = cartItems;
+    }
+  }, [cartItems, selectedCartItems, data]);
+
+  // manage addToCart button status for each product
   useEffect(() => {
     if (
       !isAllProductsLoading &&
@@ -75,7 +93,10 @@ const Cart = () => {
         );
 
         if (existingCartItem) {
-          if (existingCartItem?.quantity >= existingCartItem?.quantityInStock) {
+          if (
+            (existingCartItem?.quantity as number) >=
+            (existingCartItem?.quantityInStock as number)
+          ) {
             disabledButtons.push({ [product._id]: true });
           } else {
             disabledButtons.push({ [product._id]: false });
@@ -97,6 +118,7 @@ const Cart = () => {
     setDisabledCartButtons,
   ]);
 
+  // handle item selection in table
   const onSelectChange = (newSelectedRowKeys: React.Key[]) => {
     setSelectedRowKeys(newSelectedRowKeys);
 
@@ -106,6 +128,8 @@ const Cart = () => {
     setShowMultipleDeleteButton(items.length > 1);
   };
 
+  // actions
+  // handle delete single cart item
   const handleDeleteOne = (key: React.Key) => {
     Swal.fire({
       title: "Are you sure?",
@@ -117,14 +141,12 @@ const Cart = () => {
       confirmButtonText: "Yes, delete it!",
     }).then((result) => {
       if (result.isConfirmed) {
-        setLoading(true);
-        const selectedOne: DataType = data.find((item) => item.key === key);
+        const selectedOne = data.find((item) => item.key === key) as DataType;
 
         if (selectedOne) {
           dispatch(deleteCartItems({ selectedCartItems: [selectedOne] }));
         }
 
-        setLoading(false);
         setSelectedCartItems([]);
         Swal.fire({
           title: "Deleted!",
@@ -135,27 +157,41 @@ const Cart = () => {
     });
   };
 
+  // handle delete multiple cart items at once
   const handleDeleteMultiple = () => {
-    if (selectedCartItems.length > 0) {
-      dispatch(deleteCartItems({ selectedCartItems }));
-      setSelectedRowKeys([]);
-      setSelectedCartItems([]);
-      setShowMultipleDeleteButton(false);
-    }
+    Swal.fire({
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, delete it!",
+    }).then((result) => {
+      if (result.isConfirmed) {
+        if (selectedCartItems.length > 0) {
+          dispatch(deleteCartItems({ selectedCartItems }));
+          setSelectedRowKeys([]);
+          setSelectedCartItems([]);
+          setShowMultipleDeleteButton(false);
+        }
+        setSelectedCartItems([]);
+        Swal.fire({
+          title: "Deleted!",
+          text: "Selected items removed from your cart!",
+          icon: "success",
+        });
+      }
+    });
   };
 
+  // handle instant quantity change for each cart item
   const handleQuantityChange = (value: number, key: React.Key) => {
     const updatedItem = data.find((item) => item.key === key);
 
     if (updatedItem) {
       const maxQuantity = updatedItem.quantityInStock || 100;
       const newQuantity = Math.min(value, maxQuantity);
-
-      const updatedItemWithNewQuantity = {
-        ...updatedItem,
-        quantity: newQuantity,
-        quantityInStock: updatedItem.quantityInStock,
-      };
 
       dispatch(
         updateCartQuantity({
@@ -167,18 +203,34 @@ const Cart = () => {
     }
   };
 
+  // navigation
+  // navigate to product details page
+  const handleGoToDetails = async (key: React.Key) => {
+    const clickedOne = data.find((item) => item.key === key);
+
+    navigate(`/products/${clickedOne!._id}`);
+  };
+
+  // row and column
   const rowSelection: TableRowSelection<DataType> = {
     selectedRowKeys,
     onChange: onSelectChange,
     selections: [Table.SELECTION_ALL, Table.SELECTION_NONE],
   };
 
+  console.log({ selectedCartItems });
+
   const columns: TableColumnsType<DataType> = [
     {
       title: "Image",
       dataIndex: "image",
-      render: (image: string) => (
-        <img src={image} style={{ width: 70, height: 70 }} />
+      render: (image: string, record) => (
+        <img
+          className="transition-transform transform hover:scale-105"
+          onClick={() => handleGoToDetails(record.key)}
+          src={image}
+          style={{ width: 70, height: 70 }}
+        />
       ),
     },
     {
@@ -194,7 +246,7 @@ const Cart = () => {
             className="bg-white text-black border border-slate-300 h-8 hover:bg-rose-500 hover:text-white"
             onClick={() =>
               handleQuantityChange(
-                record.quantity - 1 > 0 ? record.quantity - 1 : 1,
+                (record.quantity ?? 1) - 1 > 0 ? (record.quantity ?? 1) - 1 : 1,
                 record.key
               )
             }
@@ -204,16 +256,18 @@ const Cart = () => {
           <Input
             className="w-12 text-center"
             min={0}
-            max={record.quantityInStock}
-            value={record.quantity}
-            onChange={(value) => handleQuantityChange(value, record.key)}
+            max={Number(record.quantityInStock)}
+            value={record.quantity ?? 0}
+            onChange={(e) =>
+              handleQuantityChange(Number(e.target.value), record.key)
+            }
           />
           <Button
             className="h-8 bg-white text-black border border-slate-300 hover:bg-rose-500 hover:text-white"
             onClick={() =>
-              handleQuantityChange(record.quantity + 1, record.key)
+              handleQuantityChange((record.quantity ?? 0) + 1, record.key)
             }
-            disabled={record.quantity === record.quantityInStock}
+            disabled={(record.quantity ?? 0) === (record.quantityInStock ?? 0)}
           >
             +
           </Button>
@@ -232,11 +286,7 @@ const Cart = () => {
           className="w-12 bg-rose-500"
           onClick={() => handleDeleteOne(record.key)}
         >
-          {loading ? (
-            "Deleting..."
-          ) : (
-            <RiDeleteBin6Fill className="w-full h-full" />
-          )}
+          <RiDeleteBin6Fill className="w-full h-full" />
         </Button>
       ),
     },
@@ -256,8 +306,25 @@ const Cart = () => {
   return (
     <div>
       {cartItems.length === 0 ? (
-        <div className="flex justify-center">
-          <img src={img} style={{ maxWidth: 350 }} alt="" />
+        <div className="flex justify-center justify-centers flex-col">
+          <div className="flex">
+            <img
+              src={img}
+              className="mx-auto mt-8 mb-4"
+              style={{ maxWidth: 350 }}
+              alt=""
+            />
+          </div>
+          <div className="text-center">
+            <h2 className="text-2xl font-semibold mb-2">Your Cart is Empty</h2>
+            <p>Looks like you haven't made order yet.</p>
+            <p
+              onClick={() => navigate("/products")}
+              className="text-blue-600 hover:cursor-pointer hover:underline font-semibold"
+            >
+              Continue to Shopping
+            </p>
+          </div>
         </div>
       ) : (
         <div>
